synth/requirements.txt
asn1crypto==1.2.0
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
conda==4.7.12
conda-package-handling==1.6.0
cryptography==2.8
cycler==0.10.0
Cython==0.29.15
idna==2.8
kiwisolver==1.1.0
matplotlib==3.1.3
numpy==1.18.1
Pillow==7.0.0
pycosat==0.6.3
pycparser==2.19
pyliblo==0.10.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
PySocks==1.7.1
python-dateutil==2.8.1
python-engineio==3.11.2
python-socketio==4.4.0
requests==2.22.0
ruamel-yaml==0.15.46
six==1.12.0
supercollider==0.0.4
torch==1.4.0
torchvision==0.5.0
tqdm==4.36.1
urllib3==1.24.2
websocket-client==0.57.0
